fix(reviews): add authorization checks for create, update, and remove operations

+++ src/reviews/reviews.controller.ts
@@ -5,24 +5,42 @@ import {
   Body,
   Patch,
   Param,
+  Req,
   Delete,
   ParseUUIDPipe,
   NotFoundException,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { ReviewsService } from './reviews.service';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
-import { ApiTags, ApiOkResponse, ApiCreatedResponse } from '@nestjs/swagger';
+import {
+  ApiTags,
+  ApiOkResponse,
+  ApiCreatedResponse,
+  ApiBearerAuth,
+} from '@nestjs/swagger';
 import { ReviewEntity } from './entities/review.entity';
+import { AuthenticatedRequest } from 'src/users/users.controller';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 @ApiTags('reviews')
 @Controller('reviews')
 export class ReviewsController {
   constructor(private readonly reviewsService: ReviewsService) {}
 
   @Post()
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiCreatedResponse({ type: ReviewEntity })
-  async create(@Body() createReviewDto: CreateReviewDto) {
-    return new ReviewEntity(await this.reviewsService.create(createReviewDto));
+  async create(
+    @Body() data: CreateReviewDto,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    if (!req.user || req.user.id !== data.user_id) {
+      throw new UnauthorizedException();
+    }
+    return new ReviewEntity(await this.reviewsService.create(data));
   }
 
   @Get()
@@ -43,19 +61,34 @@ export class ReviewsController {
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ReviewEntity })
   async update(
     @Param('id', ParseUUIDPipe) id: string,
     @Body() updateReviewDto: UpdateReviewDto,
+    @Req() req: AuthenticatedRequest,
   ) {
+    if (!req.user || req.user.id !== id) {
+      throw new UnauthorizedException();
+    }
+
     return new ReviewEntity(
       await this.reviewsService.update(id, updateReviewDto),
     );
   }
 
   @Delete(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ReviewEntity })
-  async remove(@Param('id', ParseUUIDPipe) id: string) {
+  async remove(
+    @Param('id', ParseUUIDPipe) id: string,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    if (!req.user || req.user.id !== id) {
+      throw new UnauthorizedException();
+    }
     return await this.reviewsService.remove(id);
   }
 }
